requirements.txt
fastapi==0.104.1
asyncio==3.4.3
pydantic==2.5.0
psycopg2-binary==2.9.7
chromadb==0.4.24
neo4j==5.16.0
celery==5.3.4
redis==5.0.1
llama-cpp-python==0.2.26
numpy==1.24.3
pandas==2.0.3
pytest==7.4.2
